+++ lib/td_dq/rules/rule_type.ex
@@ -8,8 +8,8 @@ defmodule TdDq.Rules.RuleType do
   @available_params_keys ["system_params", "type_params"]
 
   schema "rule_types" do
-    field :name, :string
-    field :params, :map
+    field(:name, :string)
+    field(:params, :map)
 
     timestamps()
   end
@@ -25,19 +25,24 @@ defmodule TdDq.Rules.RuleType do
 
   defp validate_params(changeset) do
     params = get_change(changeset, :params)
+
     if params == %{} or params == nil do
       changeset
     else
       params = Map.take(get_change(changeset, :params), @available_params_keys)
+
       with true <- length(Map.keys(params)) == length(@available_params_keys),
-           true <- length(Map.keys(get_change(changeset, :params))) == length(@available_params_keys) do
+           true <-
+             length(Map.keys(get_change(changeset, :params))) == length(@available_params_keys) do
         changeset
       else
         false ->
           changeset
-          |> add_error(:params, "Invalid params. Available params: #{inspect(@available_params_keys)}")
+          |> add_error(
+            :params,
+            "Invalid params. Available params: #{inspect(@available_params_keys)}"
+          )
       end
     end
   end
-
 end