+++ priv/repo/migrations/20180206094226_add_quality_control_status.exs
@@ -3,8 +3,7 @@ defmodule TdDq.Repo.Migrations.AddQualityControlStatus do
 
   def change do
     alter table("quality_controls") do
-      add :status, :string, default: "defined"
+      add(:status, :string, default: "defined")
     end
-
   end
 end