+++ priv/repo/migrations/20180314151913_add_qc_type_params.exs
@@ -3,7 +3,7 @@ defmodule TdDq.Repo.Migrations.AddQcTypeParams do
 
   def change do
     alter table(:quality_controls) do
-      add :type_params, :map, null: true
+      add(:type_params, :map, null: true)
     end
   end
 end