+++ priv/repo/migrations/20180911134050_add_parent_domains_to_rule_result.exs
@@ -4,13 +4,13 @@ defmodule TdDq.Repo.Migrations.AddParentDomainsToRuleResult do
   alias TdDq.Repo
 
   def up do
-    alter table(:rule_results), do: add :parent_domains, :text, null: true
+    alter(table(:rule_results), do: add(:parent_domains, :text, null: true))
     flush()
     SQL.query!(Repo, "update rule_results set parent_domains = ''")
-    alter table(:rule_results), do: modify :parent_domains, :text, null: false
+    alter(table(:rule_results), do: modify(:parent_domains, :text, null: false))
   end
 
   def down do
-    alter table(:rule_results), do: remove :parent_domains
+    alter(table(:rule_results), do: remove(:parent_domains))
   end
 end