+++ priv/repo/migrations/20190212145911_drop_rules_unused_fields.exs
@@ -3,8 +3,8 @@ defmodule TdDq.Repo.Migrations.DropRulesUnusedFields do
 
   def change do
     alter table(:rules) do
-      remove :type_backup
-      remove :status_backup
+      remove(:type_backup)
+      remove(:status_backup)
     end
   end
 end