+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-2144] Support ngram-search in structure name
 
+### Changed
+
+- [TD-2200] Prevalence of data structure's attributes over metadata on versions index
+
 ## [3.8.0] 2019-10-14
 
 ### Fixed