+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Added
 
 - [TD-2144] Support ngram-search in structure name
+- [TD-2159] Mapping for data field type
+
+### Changed
+
+- [TD-2200] Prevalence of data structure's attributes over metadata on versions index
 
 ## [3.8.0] 2019-10-14